requirements.txt
flask
python-dotenv
requests
groq
gunicorn
